# Update worker with rapidsnark for faster proof generation