--- a/docker/worker.Dockerfile
+++ b/docker/worker.Dockerfile
@@ -1,5 +1,6 @@
 # build circom from source for local verify
 FROM ghcr.io/eyblockchain/local-circom as builder
+FROM ghcr.io/eyblockchain/local-rapidsnark as rapidsnark
 
 FROM ubuntu:20.04
 
@@ -7,6 +8,10 @@ RUN apt-get update -y
 RUN apt-get install -y netcat curl
 RUN curl -fsSL https://deb.nodesource.com/setup_16.x | bash -
 RUN apt-get install -y nodejs gcc g++ make
+RUN apt install -y build-essential
+RUN apt-get install -y libgmp-dev
+RUN apt-get install -y libsodium-dev
+RUN apt-get install -y nasm
 
 EXPOSE 80
 
@@ -22,6 +27,7 @@ WORKDIR /app
 COPY config/default.js config/default.js
 COPY /nightfall-deployer/circuits circuits
 COPY --from=builder /app/circom/target/release/circom /app/circom
+COPY --from=rapidsnark /app/rapidsnark/build/prover /app/prover
 COPY ./worker/package.json ./worker/package-lock.json ./
 COPY ./worker/src ./src
 COPY ./worker/start-script ./start-script