## What does this implement/fix? Explain your changes.
Update worker and move from snarkjs to rapidsnark for faster proof generartion.

Performance test in a laptop with a Ryzen 7 4800HS processor with 8 cores 16 threads and 16GB of RAM.
| Prover | Deposit | Transfer | Withdraw |
|---|---|---|---|
| snarkjs | 361 ms  | 2395 ms  | 2140 ms  | 
| rapidsnark | 67 ms  |  1572 ms |  642 ms | 

## Does this close any currently open issues?
No

## What commands can I run to test the change? 
Any previous test that generates proofs in the worker. 
- npm run test-erc20-tokens
- npm run test-erc721-tokens 
- npm run test-erc1155-tokens 

## Any other comments?
Rapidsnark it's not working for ARM. We close the PR.